Version antigua/requirements.txt
# ================================================
# DEPENDENCIAS PRINCIPALES - Automatización Cotizaciones
# ================================================

# Playwright - Automatización de navegador web
playwright==1.48.0

# Python-dotenv - Manejo de variables de entorno
python-dotenv==1.0.1

# Pandas - Para manipulación de datos y creación de Excel (última versión con binarios)
pandas>=2.0.0

# NumPy - Dependencia de pandas (última versión con binarios)
numpy>=1.24.0

# OpenPyXL - Para trabajar con archivos Excel
openpyxl==3.1.2

# Pillow - Para manejo de imágenes en Excel
Pillow>=10.0.0

# psutil - Para monitoreo y gestión de procesos
psutil>=5.9.0
